Refactoring drawer and separating widgets
