+++ challenges/flutter/test_challenge/lib/app/modules/dashboard/infra/mappers/vehicle_mapper.dart
