+++ challenges/flutter/test_challenge/lib/app/modules/dashboard/presenter/widgets/dashboard_vehicle_cell/dashboard_vehicle_cell.dart
